backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
python-dotenv
boto3
tavily-python
aiohttp
httpx
